requirements.dev.txt
pytest-env
coverage
factory-boy
Faker
freezegun
pytest
pytest-cov
pytest-mock
pytest-sugar
responses
testfixtures
black
pytest-watch
mypy
pytest-mypy
pytest-flask
pytest-flask-sqlalchemy